+++ tests/testthat/test_aoi_processing.R
@@ -2,9 +2,9 @@ context('aoi processing')
 
 test_that(
   'aoi_convert functions as expected',
-  c(badx_aoi <- c(c(144.18, -40.17, 143.75, -39.57)),
-    bady_aoi <- c(c(143.75, -39.57, 144.18, -40.17)),
-    bado_aoi <- c(c(143.75, 144.18, -40.17, -39.57)),
+  c(badx_aoi <- c(144.18, -40.17, 143.75, -39.57),
+    bady_aoi <- c(143.75, -39.57, 144.18, -40.17),
+    bado_aoi <- c(143.75, 144.18, -40.17, -39.57),
     expect_error(slga:::aoi_convert(badx_aoi)),
     expect_error(slga:::aoi_convert(bady_aoi)),
     expect_error(slga:::aoi_convert(bado_aoi))
@@ -135,34 +135,80 @@ test_that(
     aoi <- structure(c(143.75, -40.17, 144.18, -39.57),
                      names = c("xmin", "ymin", "xmax", "ymax"),
                      class = "bbox", crs = sf::st_crs(4283)),
-    val1 <- slga:::validate_aoi(aoi, 'NAT'),
-    val2 <- slga:::validate_aoi(aoi, 'TAS'),
+    expect_error(validate_aoi(aoi)), # no product supplied
+    val1 <- slga:::validate_aoi(aoi, 'NAT'), # normal bbox input
+    val2 <- slga:::validate_aoi(aoi, 'TAS'), # normal bbox input
     expect_is(val1, 'bbox'),
     expect_equal(attr(val1, 'crs')$epsg, 4283),
     expect_is(val2, 'bbox'),
     expect_equal(attr(val2, 'crs')$epsg, 4283),
-    expect_error(slga:::validate_aoi(aoi, 'SA')),
+    expect_error(slga:::validate_aoi(aoi, 'SA')), # no overlap
+    # convert from simple
     aoi_simple <- c(143.75, -40.17, 144.18, -39.57),
     val3 <- slga:::validate_aoi(aoi_simple, 'NAT'),
     expect_equal(val1, val3),
+    # convert from raster extent
     aoi_raster <- raster::extent(sf::st_sf(sf::st_as_sfc(val1), 4283)),
     val4 <- slga:::validate_aoi(aoi_raster, 'NAT'),
     expect_equal(val1, val4),
+    # convert from actual raster
     library(raster),
     data('ki_surface_clay'),
     val5 <- slga:::validate_aoi(ki_surface_clay, 'NAT'),
     val6 <- slga:::validate_aoi(raster::extent(ki_surface_clay), 'NAT'),
     expect_equivalent(val5, val6),
     expect_error(slga:::validate_aoi('1', 'NAT')),
+    # convert from sfc
     val7 <- sf::st_as_sfc(aoi, crs = 4283),
     expect_equal(slga:::validate_aoi(val7, 'NAT'), val1),
+    # convert from sfg
+    expect_equal(slga:::validate_aoi(val7[[1]], 'NAT'), val1),
+    # convert from sf
     val7a <- sf::st_sf(val7),
     expect_equal(slga:::validate_aoi(val7a, 'NAT'), val1),
     val8 <- st_bbox(sf::st_transform(sf::st_as_sfc(aoi, crs = 4283), 28356)),
     val9 <- val8,
     attr(val9, 'crs')$epsg <- NA,
     expect_equal(slga:::validate_aoi(val8, 'NAT'), slga:::validate_aoi(val9, 'NAT')),
-    expect_equal(slga:::validate_aoi(val9, 'NAT'), slga:::validate_aoi(val9, 'NAT')),
-    expect_error(slga:::validate_aoi(val9))
+    expect_equal(slga:::validate_aoi(val9, 'NAT'), slga:::validate_aoi(val9, 'NAT'))
   )
 )
+
+test_that(
+  'validate_poi functions as expected',
+  c(
+    expect_error(slga:::validate_poi(c(151, 27), 'NAT')),
+    poi <- c(151, -27),
+    v1 <- slga:::validate_poi(poi, 'NAT'),
+    v2 <- slga:::validate_poi(poi, 'NAT', buff = 3),
+    expect_is(v1, 'bbox'),
+    expect_is(v2, 'bbox'),
+    expect_equal(attr(v1, 'crs')$epsg, 4283),
+    expect_equal(attr(v2, 'crs')$epsg, 4283),
+    expect_true(v1[1] < poi[1]), # v1 encloses poi
+    expect_true(v1[2] < poi[2]),
+    expect_true(v1[3] > poi[1]),
+    expect_true(v1[4] > poi[2]),
+    expect_equal(as.vector(abs(v1[1] - v1[3])), 0.0008333333), # v1 is 1x1 cell
+    expect_equal(as.vector(abs(v1[2] - v1[4])), 0.0008333333),
+    expect_true(v2[1] < v1[1]),  # v2 encloses v1
+    expect_true(v2[2] < v1[2]),
+    expect_true(v2[3] > v1[3]),
+    expect_true(v2[4] > v1[4]),
+    # poi is an sf point
+    poi_xy <- st_point(c(151, -27)),
+    v3 <- slga:::validate_poi(poi_xy, 'NAT'),
+    expect_equal(v1, v3),
+    # poi is an sfc_POINT of length 1 (e.g. from splitting an sfc)
+    poi_sfc <- sf::st_sfc(poi_xy),
+    v4 <- slga:::validate_poi(poi_sfc, 'NAT'),
+    expect_equal(v1, v4),
+    # poi is an sf object with 1 row (e.g. from splitting an sf data frame)
+    poi_sf <- sf::st_sf(poi_sfc),
+    poi_sf$id <- 1,
+    v5 <- slga:::validate_poi(poi_sf, 'NAT'),
+    expect_equal(v1, v5)
+    )
+)
+
+