+++ tests/testthat/test_helpers.R
@@ -12,3 +12,20 @@ test_that(
     expect_equal(val2, TRUE)
   )
 )
+
+test_that('filenamer is ok',
+          c(
+            expect_equal(slga_filenamer(product = 'NAT', attribute = 'CLY',
+                                       component = 'ALL', depth = 1),
+                         "NAT_CLY_ALL_000_005"))
+          )
+
+test_that('circmask is ok',
+          c(
+            v1 <- slga:::make_circ_mask(1),
+            mat <- matrix(c(NA, 0, NA, 0, 0, 0, NA, 0, NA),
+                          ncol = 3, nrow = 3, byrow = TRUE),
+            expect_equal(v1, mat),
+            expect_equal(slga:::make_circ_mask(0), as.matrix(0))
+          )
+        )